+++ src/avx2_double.h
@@ -7,7 +7,8 @@
 // Each fmadd is 8 flop.
 const uint64_t flop_per_iter = 14 * 4 * 2;
 
-__attribute__((noinline)) __m256d impl(int N, double a, double b, double c) {
+__attribute__((noinline)) __m256d impl(int64_t N, double a, double b,
+                                       double c) {
 
   __m256d A = _mm256_set1_pd(a);
   __m256d B = _mm256_set1_pd(b);
@@ -16,7 +17,7 @@ __attribute__((noinline)) __m256d impl(int N, double a, double b, double c) {
   __m256d s;
 
 #pragma clang loop unroll(disable)
-  for (int i = 0; i < N; ++i) {
+  for (int64_t i = 0; i < N; ++i) {
     __asm__("vfmadd231pd     %ymm15, %ymm14, %ymm0\n\t"
             "vfmadd231pd     %ymm15, %ymm14, %ymm1\n\t"
             "vfmadd231pd     %ymm15, %ymm14, %ymm2\n\t"
@@ -36,6 +37,6 @@ __attribute__((noinline)) __m256d impl(int N, double a, double b, double c) {
   return s;
 }
 
-void kernel(int N) { impl(N, 0.4, 1.5, 1.0); }
+void kernel(int64_t N) { impl(N, 0.4, 1.5, 1.0); }
 
 #endif
