+++ src/avx2_float.h
@@ -7,7 +7,7 @@
 // Each fmadd is 16 flop.
 const uint64_t flop_per_iter = 14 * 8 * 2;
 
-__attribute__((noinline)) __m256 impl(int N, float a, float b, float c) {
+__attribute__((noinline)) __m256 impl(int64_t N, float a, float b, float c) {
 
   __m256 A = _mm256_set1_ps(a);
   __m256 B = _mm256_set1_ps(b);
@@ -16,7 +16,7 @@ __attribute__((noinline)) __m256 impl(int N, float a, float b, float c) {
   __m256 s;
 
 #pragma clang loop unroll(disable)
-  for (int i = 0; i < N; ++i) {
+  for (int64_t i = 0; i < N; ++i) {
     __asm__("vfmadd231ps     %ymm15, %ymm14, %ymm0\n\t"
             "vfmadd231ps     %ymm15, %ymm14, %ymm1\n\t"
             "vfmadd231ps     %ymm15, %ymm14, %ymm2\n\t"
@@ -36,6 +36,6 @@ __attribute__((noinline)) __m256 impl(int N, float a, float b, float c) {
   return s;
 }
 
-void kernel(int N) { impl(N, 0.4f, 1.5f, 1.0f); }
+void kernel(int64_t N) { impl(N, 0.4f, 1.5f, 1.0f); }
 
 #endif
